Move Gsym symbolization benchmark into benches/symbolize
With this change we move the Gsym symbolization benchmark into
benches/symbolize, to co-locate it with other symbolization benchmarks
and to follow the organization of the rest of the crate.
